+++ test/EndToEndSpec.hs
@@ -1695,6 +1695,19 @@ spec = around_ (inTempDirectoryNamed "my-package") $ do
                   windows
             |]
 
+    describe "foreign libraries" $ do
+      it "Foreign Library stanza with type and options" $ do
+        [i|
+        foreign-library:
+          type: native-shared
+          options:
+            - standalone
+        |] `shouldRenderTo` (foreignLibrary "my-package" [i|
+        type: native-shared
+        options:
+            standalone
+        |])
+
     describe "executables" $ do
       it "accepts main-is as an alias for main" $ do
         [i|
@@ -2094,6 +2107,17 @@ internalLibrary name e = (package content) {packageCabalVersion = "2.0"}
   where
     content = [i|
 library #{name}
+#{indentBy 2 $ unindent e}
+  default-language: Haskell2010
+|]
+
+foreignLibrary :: String -> String -> Package
+foreignLibrary name e = package content
+  where
+    content = [i|
+foreign-library #{name}
+  other-modules:
+      Paths_my_package
 #{indentBy 2 $ unindent e}
   default-language: Haskell2010
 |]